api/requirements.txt
Flask==1.1.2
Flask-PyMongo==2.2.0
pymongo

flask-jwt-extended
requests
flask-socketio
gevent-websocket
Flask-Mail
itsdangerous

werkzeug>=1,<2
flask-session~=0.3.2
msal>=1.7,<2

pytz